refactors structure of configuration ui elements
