+++ src/main/kotlin/eu/florian_fuhrmann/musictimedtriggers/utils/configurations/annotations/VisibleChecker.kt
@@ -4,7 +4,6 @@ import eu.florian_fuhrmann.musictimedtriggers.utils.configurations.Configuration
 import kotlin.reflect.KClass
 
 @Target(AnnotationTarget.FIELD)
-@Repeatable
 annotation class VisibleWhen(val checker: KClass<out VisibleChecker<out Configuration>>, val inverted: Boolean = false)
 
 abstract class VisibleChecker<T : Configuration> {
